+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Release 3.2.1 (TO BE RELEASED)
 
+### API Changes
+
+* Convert ZMQSocket::sendMessage(...) methods to slots.
+
 ## Release 3.2.0
 
 * Improved documentation.